pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "lfw"
version = "0.1.0"
description = "Linode Firewall Policy Engine - CLI-first, SDK-hybrid"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1,<9",
    "pydantic>=2.5,<3",
    "pyyaml>=6.0,<7",
    "linode_api4>=5.0,<6",
    "netaddr>=1.2,<2",
    "maxminddb>=2.6,<3",
    "httpx>=0.27,<1",
    "python-dotenv>=1.0,<2",
    "rich>=13.7,<14",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0,<9",
    "pytest-cov>=5.0,<6",
    "respx>=0.21,<1",
]

[project.scripts]
lfw = "lfw.cli.main:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/lfw"]

[tool.pytest.ini_options]
testpaths = ["test"]
pythonpath = ["src"]
